Added fix for hiding settings tab when you are using "Name format for children" on a template to bypass the initial page creation step.
This fix checks to see if template setting: “nameContentTab” is enabled
which would add the name field to the content tab. If not, then we add
a hidden version of the name field so that new pages will save without
error.

This fix may no longer be necessary depending on the outcome of this PW
Github issue:
https://github.com/ryancramerdesign/ProcessWire/issues/1048